+++ src/problem2/fancy-form/package.json
@@ -1,18 +1,12 @@
 {
-  "name": "fancy-from",
+  "name": "fancy-form",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "antd": "^5.23.1",
-    "axios": "^1.7.9",
-    "cra-template-typescript": "1.2.0",
-    "formik": "^2.4.6",
+    "cra-template": "1.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-scripts": "5.0.1",
-    "react-select": "^5.9.0",
-    "styled-components": "^6.1.14",
-    "yup": "^1.6.1"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
