+++ src/problem2/fancy-form/src/App.test.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from './App';
