--- src/problem2/fancy-from/src/App.tsx
@@ -1,29 +1,0 @@
-import React from 'react';
-// @ts-ignore
-import logo from './logo.svg';
-import './App.css';
-import SwapForm from "./SwapForm";
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-      <SwapForm />
-    </div>
-  );
-}
-
-export default App;
